Add gitignore and rename generic variables to be specific to Temi
