+++ fleet_adapter_template/fleet_adapter_template/__init__.py
@@ -1,0 +1,2 @@
+from .fleet_adapter_temi import *
+from .TemiClientAPI import *
